feat(ui): global Ask chat drawer + context seeding + suggested questions (#266)
Community dashboard side of enterprise#266:

- Extract components/ask/ChatPanel — the transcript + input + markdown
  helpers, now the single source of truth for the /ask page AND the drawer.
- ChatDrawerProvider + ChatDrawer (Mantine right, non-blocking, keepMounted)
  + ChatLauncher, mounted globally in layout.tsx so the conversation
  survives navigation.
- "Ask about this" triggers on insight/rec list + detail pages; the panel
  sends seed_context on the first turn and shows a seed chip.
- SuggestedQuestions: LLM-generated starters with a thinking animation on
  insight/rec detail pages; a per-project "Suggested questions" Settings
  toggle (ask_suggestions_enabled, default on) gates the auto LLM calls.
- api.ts: SeedContext, AskRequest.seed_context, getAskSuggestions.
